% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imboxplot.R
\name{imboxplot}
\alias{imboxplot}
\title{imboxplot}
\usage{
data(data)

data(phenotype)

data(im)

res<-imenrich(data,im,method="zscore")

imboxplot(res,phenotype,id="group",fontsize=14)
}
\arguments{
\item{phenotype}{phenotype file from tested samples}

\item{id}{name of the phenotype column to be stratified}

\item{fontsize}{size of the font in the graph}

\item{results}{results obtained from imenrich function}
}
\description{
imboxplot
}
\examples{
data(data)
data(phenotype)
data(im)
res<-imenrich(data,im,method="zscore")
imboxplot(res,phenotype,id="group",fontsize=14)
}
