% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdes.R
\name{imdes}
\alias{imdes}
\title{imdes}
\usage{
data(res)

data(phenotype)

data(im)

res<-imenrich(data,im,method="zscore")

imdes.results<-imdes(res,phenotype$group,control="low")
}
\arguments{
\item{group}{group column from phenotype file}

\item{control}{string defining control in pheno$group data}
}
\description{
imdes
}
\examples{
 data(data)
data(pheno)
data(im)
res<-imenrich(data,im,method="zscore")
imdes.results<-imdes(res,phenotype$group,control="low")
head(imdes.results)
}
