% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imenrich.R
\name{imenrich}
\alias{imenrich}
\title{imenrich}
\usage{
data(data)

data(phenotype)

data(im)

res<-imenrich(data,im,method="zscore")

head(res[,1:6])
}
\arguments{
\item{data}{transcriptome matrix}

\item{im}{immune modulation genesets}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}
}
\description{
imenrich
}
\examples{
data(data)
data(phenotype)
data(im)
res<-imenrich(data,im,method="zscore")
head(res[,1:6])
}
