% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imheatmap.R
\name{imheatmap}
\alias{imheatmap}
\title{imheatmap}
\usage{
data(data)

data(phenotype)

data(im)

res<-imenrich(data,im,method="zscore")

imheatmap(res,phenotype,scale="none",fontsize=10)
}
\arguments{
\item{results}{results obtained from imenrich function or sub-geneset}

\item{phenotype}{phenotype file from tested samples}

\item{scale}{method of scaling "none" "row" "column"}

\item{fontsize}{size of the font on the graph}
}
\description{
imheatmap
}
\examples{
data(data)
data(phenotype)
data(im)
res<-imenrich(data,im,method="zscore")
imheatmap(res,phenotype,scale="none",fontsize=10)
}
