% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imvol.R
\name{imvol}
\alias{imvol}
\title{imvol}
\usage{
data(data)

data(phenotype)

data(im)

res<-imenrich(data,im,method="zscore")

imdes.results<-imdes(res,phenotype$group,control="low")

imvol(imdes.results,nb=9,fc=0.5,p=0.05,size=4,alpha=1)
}
\arguments{
\item{res}{results from limma deg analysis}

\item{nb}{number of gene names to write}

\item{fc}{fold change threshold}

\item{p}{p-value threshold}

\item{size}{size of the gene labels}

\item{alpha}{transparent points}
}
\description{
imvol
}
\examples{
data(data)
data(phenotype)
data(im)
res<-imenrich(data,im,method="plage")
imdes.results<-imdes(res,phenotype$group,control="low")
imvol(imdes.results,nb=9,fc=0.01,p=0.5,size=4,alpha=1)
}
